created payment controller and adjusted buttons
