+++ views/products/cart.ejs
@@ -49,8 +49,9 @@
 
                                                         <form action="/cart/out-cart" method="POST">
                                                             <td>
-                                                            <button class="icon-ctrl up-quantity"><i class="fa fa-plus-circle"></i></button>
-                                                            <input class="number-display" type="text" value ="<%=cartItems.quantity%>" size="1" name="quantity"><button class="icon-ctrl down-quantity"><i class="fa fa-minus-circle"></i></button>
+                                                                <button class="icon-ctrl down-quantity"><i class="fa fa-minus-circle"></i></button>
+                                                            <input class="number-display" type="text" value ="<%=cartItems.quantity%>" size="1" name="quantity"><button class="icon-ctrl up-quantity"><i class="fa fa-plus-circle"></i></button>
+
 
                                                     </td>
                                                     <td>
